+++ sdk/resourcemanager/storagecache/armstoragecache/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armstoragecache
 
 const (
 	moduleName    = "armstoragecache"
-	moduleVersion = "v0.3.0"
+	moduleVersion = "v0.4.0"
 )
 
 // CacheIdentityType - The type of identity used for the cache